Add role direksi, dan fix tampilan

+++ routes/web.php
@@ -27,19 +27,17 @@ Route::get('/dashboard', function () {
 Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard');
 
 
-Route::middleware(['role:admin'])->group(function () {
-    Route::get('/user', [UserController::class, 'index'])->name('user.index')->middleware('permission:users-view');
-    Route::get('/user/create', [UserController::class, 'create'])->name('user.create');
-    Route::post('/user/store-multiple', [UserController::class, 'storeMultiple'])->name('user.storeMultiple');
-    Route::get('/user/edit/{id}', [UserController::class, 'edit'])->name('user.edit');
-    Route::put('/user/update/{id}', [UserController::class, 'update'])->name('user.update');
-    Route::delete('/user/delete/{id}', [UserController::class, 'destroy'])->name('user.destroy');
+Route::middleware(['role:admin|direksi'])->group(function () {
+    Route::get('/user', [UserController::class, 'index'])->name('user.index');
+    Route::get('/user/create', [UserController::class, 'create'])->name('user.create')->middleware('permission:users-view'); //middleware untuk memberikan permission ke role yang didaftarkan di seeder
+    Route::post('/user/store-multiple', [UserController::class, 'storeMultiple'])->middleware('permission:users-view')->name('user.storeMultiple');
+    Route::get('/user/edit/{id}', [UserController::class, 'edit'])->name('user.edit')->middleware('permission:users-view');
+    Route::put('/user/update/{id}', [UserController::class, 'update'])->name('user.update')->middleware('permission:users-view');
+    Route::delete('/user/delete/{id}', [UserController::class, 'destroy'])->name('user.destroy')->middleware('permission:users-view');
+    Route::get('/user/{id}', [UserController::class, 'show'])->name('user.show');
 });
 
-
-
 Route::middleware(['role:admin|karyawan'])->group(function () {
-    Route::get('/user/{id}', [UserController::class, 'show'])->name('user.show');
     Route::get('/kpimetrics', [KpimetricController::class, 'index'])->name('kpimetrics.index');
     Route::get('/kpimetrics/show/{id}', [KpimetricController::class, 'show'])->name('kpimetrics.show');
     Route::get('/kpimetrics/create', [KpimetricController::class, 'create'])->name('kpimetrics.create');
@@ -66,7 +64,7 @@ Route::middleware(['auth'])->group(function () {
 });
 
 // Route kpirecords
-Route::middleware(['auth', 'role:admin'])->group(function () {
+Route::middleware(['auth', 'role:admin|direksi'])->group(function () {
     Route::get('/kpirecords', [\App\Http\Controllers\KpiRecordController::class, 'index'])
         ->name('kpirecords.index');
     Route::delete('/kpirecords/{id}', [\App\Http\Controllers\KpiRecordController::class, 'destroy'])
@@ -97,7 +95,7 @@ Route::middleware(['role:karyawan'])->group(function () {
     Route::get('/laporan', [HitungkpiController::class, 'laporan'])->name('laporan.index');
     Route::get('/laporan/download', [HitungkpiController::class, 'download'])->name('laporan.download');
 });
-Route::middleware(['role:admin'])->group(function () {
+Route::middleware(['role:admin|direksi'])->group(function () {
     Route::get('/laporan/admin', [App\Http\Controllers\HitungkpiController::class, 'laporanAdmin'])->name('laporan.admin');
     Route::get('/laporan/admin/download', [App\Http\Controllers\HitungkpiController::class, 'downloadLaporanAdmin'])->name('laporan.admin.download');
     Route::get('/laporan/admin/show/{id}', [App\Http\Controllers\HitungkpiController::class, 'showLaporanAdmin'])->name('laporan.admin.show');
